+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 -   Fix the disappearance of the suspicious metrics labels [#3263](https://github.com/MaibornWolff/codecharta/pull/3263)
 -   Fix UI components that look different since version 1.115.0 [#3260](https://github.com/MaibornWolff/codecharta/pull/3260)
+-   Show latest release notes in changelog dialog [#3264](https://github.com/MaibornWolff/codecharta/pull/3264)
 
 ## [1.115.0] - 2023-03-30